Cats/HTTPCats/requirements.txt
folium~=0.15.1
requests~=2.31.0
pillow~=10.2.0